Refactor packages
Most of the basic functionality between chapters 1 and 2 are the same.
