+++ stockperson/src/test/java/stockperson/db/DbTest.java
@@ -16,12 +16,12 @@
  * You should have received a copy of the GNU General Public License
  * along with StockPerson-Java. If not, see <https://www.gnu.org/licenses/>.
  */
-package stockperson.chapter1_0.db;
+package stockperson.db;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static stockperson.chapter1_0.db.Db.Db;
-import static stockperson.chapter1_0.models.Invoice.Builder.anInvoice;
-import static stockperson.chapter1_0.models.Product.ProductBuilder.aProduct;
+import static stockperson.db.Db.Db;
+import static stockperson.model.Invoice.Builder.anInvoice;
+import static stockperson.model.Product.ProductBuilder.aProduct;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
