+++ stockperson/src/test/java/stockperson/model/InvoiceLineTest.java
@@ -16,11 +16,11 @@
  * You should have received a copy of the GNU General Public License
  * along with StockPerson-Java. If not, see <https://www.gnu.org/licenses/>.
  */
-package stockperson.chapter1_0.models;
+package stockperson.model;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static stockperson.chapter1_0.models.InvoiceLine.Builder.anInvoiceLine;
-import static stockperson.chapter1_0.models.Product.ProductBuilder.aProduct;
+import static stockperson.model.InvoiceLine.Builder.anInvoiceLine;
+import static stockperson.model.Product.ProductBuilder.aProduct;
 
 import org.junit.jupiter.api.Test;
