+++ stockperson/src/test/java/stockperson/model/InvoiceTest.java
@@ -16,12 +16,12 @@
  * You should have received a copy of the GNU General Public License
  * along with StockPerson-Java. If not, see <https://www.gnu.org/licenses/>.
  */
-package stockperson.chapter1_0.models;
+package stockperson.model;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static stockperson.chapter1_0.models.Invoice.Builder.anInvoice;
-import static stockperson.chapter1_0.models.InvoiceLine.Builder.anInvoiceLine;
-import static stockperson.chapter1_0.models.Product.ProductBuilder.aProduct;
+import static stockperson.model.Invoice.Builder.anInvoice;
+import static stockperson.model.InvoiceLine.Builder.anInvoiceLine;
+import static stockperson.model.Product.ProductBuilder.aProduct;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
