+++ stockperson/src/test/java/stockperson/service/InvoiceServiceTest.java
@@ -16,15 +16,15 @@
  * You should have received a copy of the GNU General Public License
  * along with StockPerson-Java. If not, see <https://www.gnu.org/licenses/>.
  */
-package stockperson.chapter1_0.services;
+package stockperson.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static stockperson.chapter1_0.db.Db.Db;
-import static stockperson.chapter1_0.models.Invoice.Builder.anInvoice;
-import static stockperson.chapter1_0.models.InvoiceLine.Builder.anInvoiceLine;
-import static stockperson.chapter1_0.models.Product.ProductBuilder.aProduct;
-import static stockperson.chapter1_0.services.InvoiceService.invoiceFromCsv;
-import static stockperson.chapter1_0.services.InvoiceService.invoicesFromCsvFile;
+import static stockperson.db.Db.Db;
+import static stockperson.model.Invoice.Builder.anInvoice;
+import static stockperson.model.InvoiceLine.Builder.anInvoiceLine;
+import static stockperson.model.Product.ProductBuilder.aProduct;
+import static stockperson.service.InvoiceService.invoiceFromCsv;
+import static stockperson.service.InvoiceService.invoicesFromCsvFile;
 
 import java.io.File;
 import java.io.IOException;
@@ -33,8 +33,8 @@ import java.util.HashSet;
 import java.util.Set;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import stockperson.chapter1_0.db.Db;
-import stockperson.chapter1_0.models.Invoice;
+import stockperson.db.Db;
+import stockperson.model.Invoice;
 
 class InvoiceServiceTest {
